refact: moved dynamo tools
