+++ src/cores/dynamo/tools/query.ts
