+++ src/cores/dynamo/tools/scan.ts
