+++ src/cores/dynamo/tools/serializer.ts
