+++ src/cores/dynamo/tools/utils.ts
